factor consensus functions into consensus_utils, make consensus.go cleaner
